Finalized AnalogSelector logic
