+++ src/main/java/ca/team2706/frc/robot/OI.java
@@ -2,7 +2,10 @@ package ca.team2706.frc.robot;
 
 import ca.team2706.frc.robot.commands.ArcadeDriveWithJoystick;
 import ca.team2706.frc.robot.subsystems.DriveBase;
+import edu.wpi.first.wpilibj.AnalogInput;
 import edu.wpi.first.wpilibj.Joystick;
+import edu.wpi.first.wpilibj.buttons.JoystickButton;
+import edu.wpi.first.wpilibj.command.PrintCommand;
 
 
 /**
@@ -60,7 +63,7 @@ public class OI {
         // Set subsystem default commands
         DriveBase.getInstance().setDefaultCommand(
                 new ArcadeDriveWithJoystick(driverStick, 5, true, 4, false));
-    }
+        }
 
     /**
      * Gets the driver joystick.
