+++ src/main/java/ca/team2706/frc/robot/Robot.java
@@ -1,17 +1,22 @@
 package ca.team2706.frc.robot;
 
 import ca.team2706.frc.robot.config.Config;
+import ca.team2706.frc.robot.sensors.AnalogSelector;
 import ca.team2706.frc.robot.subsystems.Bling;
 import ca.team2706.frc.robot.subsystems.DriveBase;
 import ca.team2706.frc.robot.subsystems.SensorExtras;
 import edu.wpi.first.cameraserver.CameraServer;
+import edu.wpi.first.wpilibj.AnalogInput;
 import edu.wpi.first.wpilibj.RobotBase;
 import edu.wpi.first.wpilibj.TimedRobot;
+import edu.wpi.first.wpilibj.command.Command;
+import edu.wpi.first.wpilibj.command.PrintCommand;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
 
 import java.util.ArrayList;
 import java.util.function.Consumer;
+
 
 /**
  * Main Robot class
@@ -70,6 +75,8 @@ public class Robot extends TimedRobot {
         Scheduler.getInstance().run();
     }
 
+    private Command[] commands = {};
+
     /**
      * Caled at the beginning of autonomous.
      */
@@ -77,6 +84,17 @@ public class Robot extends TimedRobot {
     public void autonomousInit() {
         // Iterate through each of the state-change listeners and call them.
         onStateChange(RobotState.AUTONOMOUS);
+
+        // The index based the voltage of the selector
+        int index = DriveBase.getInstance().getAnalogSelectorIndex();
+
+        // Check to see if the command exists in the desired index
+        if (DriveBase.getInstance().getAnalogSelectorIndex() < commands.length && commands[index] != null) {
+            commands[DriveBase.getInstance().getAnalogSelectorIndex()].start();
+        } else if (commands.length > 0 && commands[0] != null) {
+            commands[0].start();
+        }
+
     }
 
     /**
@@ -121,6 +139,7 @@ public class Robot extends TimedRobot {
      */
     @Override
     public void testPeriodic() {
+
     }
 
     /**
